requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
cachetools==5.3.1
celery==5.3.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
gunicorn==21.2.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
imgkit==1.2.3
importlib-metadata==6.8.0
kombu==5.3.1
multidict==6.0.4
mysqlclient==2.2.0
numpy==1.25.2
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pdfkit==1.0.0
Pillow==10.0.0
prompt-toolkit==3.0.39
protobuf==4.24.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyee==8.2.2
PyJWT==1.7.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.6.3
websockets==10.4
yarl==1.9.2
zipp==3.17.0
